Update grub2 with argon2 on backup.nix
This uses the latest grub2-git commit, plus patches in AUR's grub-improved-luks2-git. Notice `type-fix.patch` is due to outdated gnulib, but that can't be on latest commit since it breaks build; the commit I used in let/in makes it work.
This is done in a way to keep maximum compatibility with nixpkgs's grub2 package.